+++ pkg/shell/gatecheck.go
@@ -1,0 +1,25 @@
+package shell
+
+import "os/exec"
+
+// GatecheckList will print a summarized view of a a report
+//
+// Requirement: supported report from STDIN
+//
+// Output: table to STDOUT
+func GatecheckList(options ...OptionFunc) ExitCode {
+	o := newOptions(options...)
+	cmd := exec.Command("gatecheck", "list", "--input-type", o.reportType)
+	return run(cmd, o)
+}
+
+// GatecheckListAll will print a summarized view of a a report with EPSS scores
+//
+// Requirement: supported report from STDIN
+//
+// Output: table to STDOUT
+func GatecheckListAll(options ...OptionFunc) ExitCode {
+	o := newOptions(options...)
+	cmd := exec.Command("gatecheck", "list", "--all", "--input-type", o.reportType)
+	return run(cmd, o)
+}